% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimvis.R
\name{.checkDistMat}
\alias{.checkDistMat}
\title{Check that the mapping object actually corresponds to points in a_data and b_data}
\usage{
.checkDistMat(a_data, b_data, dist_mat)
}
\description{
Check that the mapping object actually corresponds to points in a_data and b_data
}
