% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimvis.R
\name{.prepareMilo}
\alias{.prepareMilo}
\title{Make sure data is in correct format for scimvis
e.g.
each point has id attribute
each point has x_coord and y_coord values}
\usage{
.prepareMilo(
  a_milo,
  b_milo,
  dist_mat,
  dimred,
  N_max,
  colour_by,
  palette,
  a_title,
  b_title
)
}
\description{
Make sure data is in correct format for scimvis
e.g.
each point has id attribute
each point has x_coord and y_coord values
}
