% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimvis.R
\name{getTopN}
\alias{getTopN}
\title{Extract top N cells from a similarity matrix.}
\usage{
getTopN(sim_mat, N, margin = 1)
}
\description{
Given a matrix of similarity values, extract the top N columns for each row
}
