% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{highlightUMAP}
\alias{highlightUMAP}
\title{Highlights cells in a dimensionality reduction plot.}
\usage{
highlightUMAP(
  A_sce,
  dimred,
  colour_by,
  subset,
  palette = NULL,
  point_size = NULL,
  point_shape = 16,
  alpha_low = 0.2,
  legend_point_size = 6
)
}
\description{
Highlights cells in a dimensionality reduction plot.
}
