% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCellMappings}
\alias{plotCellMappings}
\title{Plot related cells across low-dimensional embeddings}
\usage{
plotCellMappings(
  A_sce,
  B_sce,
  sim_mat,
  dimred = "FA",
  colour_by = "AO_celltypes",
  subset_by = "AO_celltypes",
  subset = c("Hypo-Retina-like region"),
  colour_sim = FALSE,
  plot_both = TRUE,
  offset = c(10, 0),
  palette = NULL,
  colour_map = "viridis",
  point_size = NULL,
  point_shape = 16,
  alpha_low = 0.2,
  legend_point_size = 6
)
}
\description{
Plot related cells across low-dimensional embeddings
}
