% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMappingsOnRef}
\alias{plotMappingsOnRef}
\title{Colour cells according to their closest related cell in another dataset.}
\usage{
plotMappingsOnRef(
  A_sce,
  B_sce,
  sim_mat,
  dimred = "UMAP",
  colour_by = "AO_celltypes",
  subset_by = "AO_celltypes",
  subset = c("Hypo-Retina-like region"),
  colour_sim = FALSE,
  plot_both = TRUE,
  offset = c(10, 0),
  palette = NULL,
  colour_map = "viridis",
  point_size = NULL,
  point_shape = 16,
  alpha_low = 0.2,
  legend_point_size = 6
)
}
\description{
Colour cells according to their closest related cell in another dataset.
}
