% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimvis.R
\name{prepareDataSCE}
\alias{prepareDataSCE}
\title{Extracts SCE dimred and colData info}
\usage{
prepareDataSCE(data, dimred, colour_by, palette)
}
\description{
Extracts metadata and embedding coordinates corresponding to each
cell/neighbourhood from a SCE/Milo object
}
