% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimvis.R
\name{prepareNhoodData}
\alias{prepareNhoodData}
\title{Extract neighbourhood information}
\usage{
prepareNhoodData(milo, dimred, colour_by, palette)
}
\description{
Extracts metadata and embedding coordinates corresponding to each
neighbourhood from a Milo object
}
