% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scimvis.R
\name{scimvis}
\alias{scimvis}
\title{Display a scimvis widget.}
\usage{
scimvis(
  a_data,
  b_data,
  dist_mat,
  dimred = "UMAP",
  N_max = 5,
  colour_by,
  palette = NULL,
  point_size = 2,
  a_title = "Dataset A",
  b_title = "Dataset B",
  opacity_low = 0.1,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{dist_mat}{A matrix of distance/similarity values between points of
a_data and b_data.}
}
\description{
Display a scimvis widget.
}
\examples{
scimvis(r_milo, m_milo, nhood_sim, colour_by = "celltype",
palette = scrabbitr::getCelltypeColours())


}
